:books: Add custom navigation styling docs and extract NavItem component
